services/generation-server/generation_server/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
Pillow
google-genai
sentry-sdk
supabase>=2.5.0
